+++ src/core/config.module.ts
@@ -1,9 +1,14 @@
+import { readFileSync } from 'node:fs'
+
 import { ConfigModule } from '@nestjs/config'
 import * as Joi from 'joi'
 
 const { env } = process
 
 export const config = () => ({
+  // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+  version: String(JSON.parse(readFileSync('package.json', 'utf8')).version),
+
   node: {
     env: env.NODE_ENV,
   },